Fix stroke drawing glitch.
This was happening because stroke points were ending up out of order. I tracked it down to this for
each loop, which packs a sparse array into a dense one. A for-each on a sparse array is supposed to
skip over the holes. It still does in Flash 11, but for whatever reason iteration happens out of
order (!?).

Maybe the behavior of for-each on a sparse array is actually non-standard. Or maybe _points was
being typed to a dynamic Object and it was iterating over the keys.


git-svn-id: svn://localhost/tmp/repo1/trunk@209 2a70908b-3aae-40ec-bb78-60ac196837c9